Restore prometheus
